feat: 5.0.2 (foc sheet, fate-condensed pt-br, darker index cards color) (#299)
- feat: update the Fate of Cthulhu template to include the corruption track
- feat: fate condensed is now fully available in pt-br
- fix: index card colors in dark mode have been adjusted so that they are easier on the eye
- feat: the index card color picker now offers a set of 12 colors to help you organize your cards better
- fix: there is now a confirmation pop-up if you try to delete an index card

+++ lib/components/IndexCard/IndexCard.tsx
@@ -53,6 +53,7 @@ import { ConditionalWrapper } from "../ConditionalWrapper/ConditionalWrapper";
 import { ContentEditable } from "../ContentEditable/ContentEditable";
 import { IndexCardSkills } from "./domains/IndexCardSkills";
 import { useIndexCard } from "./hooks/useIndexCard";
+import { IndexCardColor, IndexCardColorTypes } from "./IndexCardColor";
 
 export const IndexCard: React.FC<
   {
@@ -102,7 +103,11 @@ export const IndexCard: React.FC<
     xs: 1,
   });
 
-  const paper = useTextColors(indexCardManager.state.indexCard.color);
+  const paper = useTextColors(
+    theme.palette.type === "light"
+      ? indexCardManager.state.indexCard.color
+      : darken(indexCardManager.state.indexCard.color, 0.5)
+  );
 
   const defaultButtonTheme = useThemeFromColor(paper.primary);
 
@@ -374,7 +379,12 @@ export const IndexCard: React.FC<
                     size="small"
                     data-cy={`${props["data-cy"]}.remove`}
                     onClick={() => {
-                      props.onRemove();
+                      const confirmed = confirm(
+                        t("index-card.remove-confirmation")
+                      );
+                      if (confirmed) {
+                        props.onRemove();
+                      }
                     }}
                   >
                     <DeleteIcon htmlColor={paper.primary} />
@@ -841,13 +851,51 @@ function IndexCardColorPicker(props: {
     onChange: props.onChange,
     delay: 75,
   });
+
   return (
-    <ChromePicker
-      color={color}
-      disableAlpha
-      onChange={(color) => {
-        return setColor(color.hex);
-      }}
-    />
+    <>
+      <ChromePicker
+        color={color}
+        disableAlpha
+        onChange={(color) => {
+          return setColor(color.hex);
+        }}
+        styles={{
+          default: {
+            picker: {
+              boxShadow: "none",
+            },
+          },
+        }}
+      />
+      <Box pb=".5rem" bgcolor="white" width="225px">
+        <Grid container justify="center" spacing={1}>
+          {Object.keys(IndexCardColor).map((colorName) => {
+            const color = IndexCardColor[colorName as IndexCardColorTypes];
+            return (
+              <Grid item key={color}>
+                <IconButton
+                  size="small"
+                  onClick={() => {
+                    setColor(color);
+                    props.onChange(color);
+                  }}
+                >
+                  <Box
+                    className={css({
+                      width: "1.5rem",
+                      height: "1.5rem",
+                      background: color,
+                      borderRadius: "50%",
+                      border: "1px solid #e0e0e0",
+                    })}
+                  />
+                </IconButton>
+              </Grid>
+            );
+          })}
+        </Grid>
+      </Box>
+    </>
   );
 }
